servers/mcp-vectorstore/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-vectorstore"
version = "0.1.0"
description = "MCP server for vector database operations in RAG pipelines"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "RAG ChatBot Team"}
]
keywords = ["mcp", "vector-database", "rag", "embeddings", "chromadb", "faiss"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp>=0.1.0",
    "chromadb>=0.4.0",
]

[project.optional-dependencies]
all = [
    "mcp-vectorstore[faiss,dev]",
]
faiss = [
    "faiss-cpu>=1.7.0",
    "numpy>=1.24.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/rag-chatbot-mcp"
Repository = "https://github.com/yourusername/rag-chatbot-mcp"
Documentation = "https://github.com/yourusername/rag-chatbot-mcp/tree/main/servers/mcp-vectorstore"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --strict-markers"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
